% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abund_generator.R
\name{abund_generator_exp}
\alias{abund_generator_exp}
\alias{abund_generator_rlnorm}
\title{Determine yearly abundance index}
\usage{
abund_generator_exp(years, ...)

abund_generator_rlnorm(years, ...)
}
\arguments{
\item{years}{Years to generate abundance index for as a numeric vector.
For \code{abund_generator_exp()}, exponential growth/decline occurs between years,
so c(1,2,10) will give different results than c(1,2,3). For \code{abund_generator_rlnorm()},
each year's value is IID.}

\item{...}{For \code{abund_generator_exp()}, needs \code{growth.rate} (yearly population growth rate) and \code{init.size} (abundance index in the first year).
For our purposes, recommend a negative growth rate. For \code{abund_generator.rlnorm},
needs \code{meanlog} and \code{sdlog}, the mean and standard deviation of the normal distribution on a log scale.}
}
\value{
Numeric vector of abundance indices for each year.
}
\description{
From a sequence of years, generates abundance indices to feed into \code{\link[=activity_generator]{activity_generator()}}.
Currently two methods are implemented. \code{abund_generator_exp()} produces deterministic
exponential decline or growth from an initial abundnace. \code{abund_generator_rlnorm()} produces
abundances from a lognormal distribution.
}
\examples{
out = abund_generator_exp(years = 1:10, growth.rate = -0.1, init.size=500)
plot(x = 1:10, y = out)
out = abund_generator_rlnorm(years = 1:10, meanlog = 6, sdlog = 2)
plot(x = 1:10, y = out)
}
