% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_sampler.R
\name{activity_sampler}
\alias{activity_sampler}
\alias{activity_samp_pois}
\alias{activity_samp_nb}
\alias{activity_samp_zinb}
\title{Generate sample counts from expected activity}
\usage{
activity_sampler(activity.vec, sample.type, ...)

activity_samp_pois(activity.vec)

activity_samp_nb(activity.vec, ...)

activity_samp_zinb(activity.vec, ...)
}
\arguments{
\item{activity.vec}{Vector of the expected activity (see \code{?activity_gen} to generate these).}

\item{sample.type}{Name of sampling distribution. Currently supported: Poisson (\code{"pois"}),
negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}). "nb" and "zinb"
will require additional parameters.}

\item{...}{additional arguments for the sample scheme. See details}
}
\value{
Vector of simulated counts
}
\description{
Generate sample counts from expected activity
}
\details{
For Poisson sampling, no additional parameters are needed.

For the negative binomial distribution, additional argument \code{theta} must be provided.
This uses the rnegbin() function of the \code{MASS} package, with variance of \code{mu + mu^2/theta}.

For the zero-inflated negative binomial, \code{theta} and \code{pzero} must be provided. \code{pzero}
is the probability of a zero before the negative binomial is even calculated.
Note that for this distribution, the mean of generated counts \emph{will be lower than
the expected abundance based on the activity curve}.
}
\examples{

library(ggplot2)
library(tidyverse)
dat = expand.grid(years = seq(1:10),
                 doy = seq(100,150, by = 7))
abund.merge = data.frame(years = unique(dat$years),
                        abund = abund_generator(unique(dat$years),
                        abund.type = "exp",
                        growth.rate = -0.1, init.size=500)
)
dat = merge(dat, abund.merge)
dat$act = activity_gen(abund.vec = dat$abund,
                      doy = dat$doy,
                      activity.type = "Gauss",
                      act.mean = 130,
                      act.sd = 10)
dat$count = activity_sampler(dat$act,
                            sample.type = "pois")

dat.detail = expand.grid(years = unique(dat$years),
                        doy = seq(50,250, by = .1))
dat.detail = merge(dat.detail, abund.merge)
dat.detail$act = activity_gen(abund.vec = dat.detail$abund,
                             doy = dat.detail$doy,
                             activity.type = "Gauss",
                             act.mean = 130,
                             act.sd = 10)

#ggplot(data = dat, aes(x = doy, y = act))+
 #geom_point(shape = 1, col = 'blue')+
 #geom_point(aes(y = count))+
 #geom_line(data = dat.detail)+
 #facet_wrap(~years)

}
