% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_gam_sim.R
\name{butterfly_gam_sim}
\alias{butterfly_gam_sim}
\title{Title}
\usage{
butterfly_gam_sim(
  sim.name,
  path,
  nsims,
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  sim.parms,
  gam.args,
  nobs.min,
  nnzero.min,
  nyear.min,
  bound.reasonable.rel = F,
  bound.reasonable.abs = F,
  append = FALSE,
  timeseries.do = TRUE,
  gam.do = TRUE
)
}
\arguments{
\item{gam.do}{}
}
\description{
Title
}
\examples{
#  path.res = "G:/Dropbox/academia/research projects/butterfly-gam-sims/4_res"
#  gam.args = data.frame(doy.smooth = c("cc", "cc", "cc"),
#                       doy.knots = c(5,3, 5),
#                       years.smooth = c("cr", "cr", "cr"),
#                       anchor.flag = c(TRUE, TRUE, FALSE),
#                       anchor.dist = c(10, 10, 10))
# sim.parms = list(growth.rate = -0.12,
#                  init.size = 500,
#                  act.mean = 130,
#                  act.sd = 15,
#                  theta = 5)
# res = butterfly_gam_sim(sim.name = "full-kaboodle",
#                   path = path.res,
#                   nsims = 10,
#                   years = 1990:2010,
#                   doy.samples = seq(0, 365, by = .1),
#                   abund.type = "exp",
#                   activity.type = "gauss",
#                   sample.type = "pois",
#                   sim.parms = sim.parms,
#                   gam.args = gam.args,
#                   nobs.min = 5,
#                   nnzero.min = 5,
#                   nyear.min = 3)
}
