% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_gam_sim.R
\name{butterfly_load}
\alias{butterfly_load}
\title{Load time series, yearly gam estimates, and trend information from simulated parameters}
\usage{
butterfly_load(sim.name, path)
}
\arguments{
\item{sim.name}{Name for saving simulations of this combination of parameters}

\item{path}{File path to save simulated data and results into}
}
\value{
List with all relevant data from butterfly simulations.
\verb{$timeseries} gives the simulated data as if reported by butterfly survey teams (\code{years}, \code{doy}, \code{count}).
Additionally, because we simulated the data, we know what the TRUE underlying behavior was. The \code{act} column
shows the actual activity for each day based on the activity curve (\code{count} is a random sampling with that mean), and the \verb{*.true} columns
show the underlying behavior for that year based on the actual activity curve:
the true total abundance for that year as determined by \code{abund.type} and associated parameters (\code{abund.true}),
the true onset (0.1 quantile, \code{onset.true}), median (\code{median.true}), end (0.9 quantile, \code{end.true}),
and flight period (days between 0.1 and 0.9 quantiles, \code{fp.true}) for that year, as determined by the
activity curve. The \code{sim.id} column differentiates different instances of the simulation (IE different
simulated data sets using the same parameters).

\verb{$yearly.metrics} contains summary information for each year of simualted data for each simulation,
as well as the estimated abundance index and phenology metrics based on the gam fit. \code{sim.id} distinguishes
between different instances of the simulation (as in \verb{$timeseries}), and \code{gam.id} distinguishes between
different sets of gam-fitting parameters, so that we can compare how different approaches to fitting
produce different yearly estimates. \code{years} gives each individual year, \code{n} goves the total number
of simulated data points in that year, \code{nzero} gives the number of data points with 0 counts (high nzero
can lead to unreliable gam fits), and \code{abund}, \code{onset}, \code{median}, \code{end} and \code{fp} are estimated metrics from
the gam predictions of the activity curve for that year. The \code{boundary.reasonable.rel} and \code{boundary.reasonable.abs}
are attempts at automatically detecting poor gam fits (in the extreme, predicted activity curves that are U-shaped).
In general, \code{FALSE} indcates an unreasonable fit; see \code{gam_summarizer()} for details.

\verb{$trend} contains the estimated trends in abundance and phenology, the true trends in abundance and phenology, a ton of additional summary information
including gam-fitting parameters. \code{growth.rate} is the slope of the linear trend in log abundance index across years,
which is generally a good approximation for population growth rate (assuming similar observation
effort and detectability across years). \code{median}, \code{onset} \code{end}, and \code{fp} are the linear trends in
estimated phenology metrics across years. \code{nyears} shows the number of years that passed the censoring filter
and so were used in estimating these trends, and \code{passed.filtering} shows whether \code{nyears} was
above the minimum number of years defined by the \code{nyear.min} argument (if \code{passed.filtering} is \code{FALSE},
estimated trends will all be \code{NA}). \code{nyear.original} gives the number of years originally simulated,
and \code{nyear.min}, \code{nyobs.min}, and \code{nnzero.min} show the filtering criterion parameters used.
\code{bound.reasonable.rel} and \code{bound.reasonable.abs} show whether these two metrics were used
to further filter data-fitting. \code{sim.id} and \code{gam.id} identify the timeseries instance and the
gam arguments instance for this row of data. \verb{true.*} columns show the TRUE trends in log abundance index,
median, onset, end, and fp. Note that because \code{true.fp} is calculated numberically, it is slightly off from zero.
Currently simulations are designed to generate a constant phenology through time (with abundance shifting and
variable sampling noise), so the true trends in phenology should all be ~0. \verb{unreasonable.*} give the
number of years that were identified as unreasonable based on relative thresholds, absolute thresholds, or either.
These are calculated independently of whether these classifications were used as an additional filter
before calculating the trend, and may be useful for identifying gam parameterizations that are more
or less effective. The final columns are the gam parameters used: the type of smooth (cylic or cubic regression)
used for smoothing in the day-of-year direction (\code{doy.smooth}) and the year direction (\code{years.smooth}),
the number of knots used in the doy direction (the years direction always uses one knot per year), and
whether or not "anchoring zeroes" should be added outside of the observed data (\code{anchor.flag}) and
how far out to place them (\code{anchor.dist})

\verb{$parm.timeseries} gives the parameters used to generate each time series, and \verb{$parms.gam} gives the
parameters used for fitting the gams to each time series. Note that when \code{sim.parms} is given 2-value vectors for
one or more list entry, each time series has a different value for that entry drawn from a uniform
distribution with the specified values as bounds. When working with \verb{$timeseries} and
\verb{$yearly.metrics}, \code{sim.id} and \code{gam.id} can be used to match individual simulations or gam-fitting parameters
to the associated timeseries or yearly estimates.
}
\description{
Load time series, yearly gam estimates, and trend information from simulated parameters
}
\examples{
#  path.res = "G:/Dropbox/academia/research projects/butterfly-gam-sims/4_res"
#  gam.args = data.frame(doy.smooth = c("cc", "cc", "cc"),
#                       doy.knots = c(5,3, 5),
#                       years.smooth = c("cr", "cr", "cr"),
#                       anchor.flag = c(TRUE, TRUE, FALSE),
#                       anchor.dist = c(10, 10, 10))
# sim.parms = list(growth.rate = -0.12,
#                  init.size = 500,
#                  act.mean = 130,
#                  act.sd = 15,
#                  theta = 5)
# butterfly_gam_sim(sim.name = "full-kaboodle",
#                   path = path.res,
#                   nsims = 10,
#                   years = 1990:2010,
#                   doy.samples = seq(0, 365, by = .1),
#                   abund.type = "exp",
#                   activity.type = "gauss",
#                   sample.type = "pois",
#                   sim.parms = sim.parms,
#                   gam.args = gam.args,
#                   nobs.min = 5,
#                   nnzero.min = 5,
#                   nyear.min = 3)
# res = butterfly_load(sim.name = "full-kaboodle",
#                      path = path.res)
}
