% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy_selector.R
\name{doy_selector}
\alias{doy_selector}
\title{Identify reasonable first and last sample days from activity curve}
\usage{
doy_selector(height.threshold, activity.type, make.plot = TRUE, ...)
}
\arguments{
\item{height.threshold}{Identifies the minimum scaled activity for "reasonable sampling".
This is based on the maximum of the activity curve (so independent of actual abundance). Reasonable values
seem to be between 0.05 and 0.001, but use \code{make.plot} to explore.}

\item{activity.type}{What functional form to use for the underlying activity curves?
Currently supports the Zonneveld model (\code{"zon"}), #' a simple gaussian activity curve (\code{"gauss"}),
and a bivoltine activity curve based on two Gaussian curves (\code{"bivolt"}).}

\item{make.plot}{If \code{TRUE}, provide a plot of the activity curve and the proposed first and last sampling days.}

\item{...}{further arguments passed to the activity curves. See details.}
}
\value{
named vector of the first and last day of year to be used
}
\description{
Use parameterized activity curve to identify reasonable first and last days of sampling based on threshold
of activity (relative to maximum activity). This captures our expectation that sampling is
based on knowledgeable practitioners who time their sampling based on butterfly activity.
}
\examples{
doy_selector(height.threshold = 0.004, activity.type = "gauss", act.mean = 100, act.sd = 14)
}
