% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_fitall.R
\name{gam_fitall_gen}
\alias{gam_fitall_gen}
\title{Helper function to run gam_fitall and save results in a systematic way.}
\usage{
gam_fitall_gen(sim.name, path, timeseries, gam.args, append = FALSE)
}
\arguments{
\item{sim.name}{Name of this simulation instance.}

\item{path}{Filepath to save results into. Likely to remain constant across different calls of \code{timeseries_gen()}}

\item{timeseries}{Collection of timeseries, including \verb{$years}, \verb{$doy}, \verb{$count}, and \verb{$sim.id}.
Typically the \verb{$timeseries} output of \code{timeseries_sim()} or \code{timeseries_load()}. If speed is
of particular concern, can pass the trimmed data frame (\code{timeseries_sim()} generates
many more columns.)}

\item{gam.args}{Data frame, each row corresponding to a different set of gam method arguments.
Must include columns \verb{$doy.smooth}, \verb{$doy.knots},\verb{$years.smooth},\verb{$anchor.flag}, \verb{$anchor.dist}.
See \code{gam_fitter()} for details. \verb{$anchor.dist} can be any int when \verb{$anchor.flag==FALSE};
I recommend using a large negative number to make it easy to catch mistakes.}

\item{append}{Defaults to FALSE. If true, add results to existing gam fits csvs instead
of replacing them.}
}
\value{
Saves csv of yearly estimates and gam-fitting parms
}
\description{
Path and sim name should match those used in companion \code{timeseries_gen()} call. Instead
of returning fitted gam objects, saves them to file.
}
\examples{
# path.res = "G:/Dropbox/academia/research projects/butterfly-gam-sims/4_res"
# parms = list(growth.rate = c(1,-1),
#              init.size = 500,
#              act.mean = 130,
#              act.sd = 15)
# timeseries_gen(sim.name = "testsim2",
#                path = path.res,
#                nsims = 5,
#                years = 1990:2000,
#                doy.samples = seq(100,160, by = 10),
#                abund.type = "exp",
#                activity.type = "gauss",
#                sample.type = "pois",
#                sim.parms = parms)
# out = timeseries_load(sim.name = "testsim2",
#                       path = path.res)
# gam.args = data.frame(doy.smooth = c("cc", "cc"),
#                       doy.knots = c(5,3),
#                       years.smooth = c("cr", "cr"),
#                       anchor.flag = c(TRUE, TRUE),
#                       anchor.dist = c(10, 10))
# out.fit = gam_fitall(out$timeseries, gam.args)
# out.fit = gam_fitall_gen(sim.name = "testsim2",
#                          path = path.res,
#                          out$timeseries,
#                          gam.args)
# temp = gam_fitall_load(sim.name = "testsim2",
#                        path = path.res)
}
