% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_fitall.R
\name{gam_fitall_load}
\alias{gam_fitall_load}
\title{Load previously saved gam fits}
\usage{
gam_fitall_load(sim.name, path)
}
\value{
list of two data frames. \verb{$yearly.estimates} contains the metrics estimated
for each year of each simulation for each gam method (\verb{$sim.id} and \verb{$gam.id}). \verb{$gam.args}
gives the associated gam arguments that map to each gam.id in \verb{$yearly.estimates}.
}
\description{
Helper function for using \code{gam_fitall_gen()}.
}
\examples{
# path.res = "G:/Dropbox/academia/research projects/butterfly-gam-sims/4_res"
# parms = list(growth.rate = c(1,-1),
#              init.size = 500,
#              act.mean = 130,
#              act.sd = 15)
# timeseries_gen(sim.name = "testsim2",
#                path = path.res,
#                nsims = 5,
#                years = 1990:2000,
#                doy.samples = seq(100,160, by = 10),
#                abund.type = "exp",
#                activity.type = "gauss",
#                sample.type = "pois",
#                sim.parms = parms)
# out = timeseries_load(sim.name = "testsim2",
#                       path = path.res)
# gam.args = data.frame(doy.smooth = c("cc", "cc"),
#                       doy.knots = c(5,3),
#                       years.smooth = c("cr", "cr"),
#                       anchor.flag = c(TRUE, TRUE),
#                       anchor.dist = c(10, 10))
# out.fit = gam_fitall(out$timeseries, gam.args)
# out.fit = gam_fitall_gen(sim.name = "testsim2",
#                          path = path.res,
#                          out$timeseries,
#                          gam.args)
# temp = gam_fitall_load(sim.name = "testsim2",
#                        path = path.res)
}
