% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_fitall.R
\name{gam_fitall_par}
\alias{gam_fitall_par}
\title{Parallelized version of gam_fitall}
\usage{
gam_fitall_par(timeseries, gam.args, nthreads)
}
\arguments{
\item{timeseries}{Collection of timeseries, including \verb{$years}, \verb{$doy}, \verb{$count}, and \verb{$sim.id}.
Typically the \verb{$timeseries} output of \code{timeseries_sim()} or \code{timeseries_load()}. If speed is
of particular concern, can pass the trimmed data frame (\code{timeseries_sim()} generates
many more columns.)}

\item{gam.args}{Data frame, each row corresponding to a different set of gam method arguments.
Must include columns \verb{$doy.smooth}, \verb{$doy.knots},\verb{$years.smooth},\verb{$anchor.flag}, \verb{$anchor.dist}.
See \code{gam_fitter()} for details. \verb{$anchor.dist} can be any int when \verb{$anchor.flag==FALSE};
I recommend using a large negative number to make it easy to catch mistakes.}

\item{nthreads}{Number of threads to use for parallel processing. Note that function includes a check of total threads available on
the machine, and will not use more than that.}
}
\value{
list of two data frames. \verb{$yearly.estimates} contains the metrics estimated
for each year of each simulation for each gam method (\verb{$sim.id} and \verb{$gam.id}). \verb{$gam.args}
gives the associated gam arguments that map to each gam.id in \verb{$yearly.estimates}.
}
\description{
Each separate timeseries (identified with \verb{$sim.id}) will be fit separately, once for each
gam method. GAM methods are defined with separate rows of the \verb{$gam.args} arguments.
}
