% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{timeseries_examples}
\alias{timeseries_examples}
\title{Plot example timeseries}
\usage{
timeseries_examples(
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  xlim = c(0, 365),
  ...
)
}
\arguments{
\item{years}{Numeric vector of years for which which to simulate data}

\item{doy.samples}{Numeric vector of day of years for which to simulate censusing.}

\item{abund.type}{What type of variation in abundance to model. Currently supported: deterministic
exponential growth (\code{"exp"}), log normal random draws (\code{"rlnorm"}).}

\item{activity.type}{What functional form to use for the underlying activity curves?
Currently supports the Zonneveld model (\code{"zon"}), #' a simple gaussian activity curve (\code{"gauss"}),
and a bivoltine activity curve based on two Gaussian curves (\code{"bivolt"}).}

\item{sample.type}{Distribution to use when simulating censuses. The mean of this distribution
is determined by the underlying activity curve, and this distribution (and associated parameters)
determine the "sampling error" (doesn't need to strictly represent error in simulated observers).
Currently supported: Poisson (\code{"pois"}), negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}).
"nb" and "zinb" will require additional parameters (see details)}

\item{xlim}{vector of limits for plotting}

\item{...}{additional arguments for the various timeseries generation functions. See details.}
}
\value{
ggplot object
}
\description{
Plot example timeseries
}
\details{
For all the additional arguments needed for the various component functions,
see \code{?abund_generator}, \code{activity_gen}, and \code{activity_sampler}. Unfortunately it is not
possible to import details from multiple functions.
}
\examples{
timeseries_examples(years = 1990:2000,
  doy.samples = seq(105,160, by = 7),
  abund.type = "exp",
  activity.type = "gauss",
  sample.type = "pois",
  growth.rate = -0.12,
  init.size = 500,
  act.mean = 130,
  act.sd = 15)
}
