% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_generator.R
\name{timeseries_generator}
\alias{timeseries_generator}
\title{Generate time series}
\usage{
timeseries_generator(
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  ...
)
}
\arguments{
\item{years}{Numeric vector of years for which which to simulate data}

\item{doy.samples}{Numeric vector of day of years for which to simulate censusing.}

\item{abund.type}{What type of variation in abundance to model. Currently supported: deterministic
exponential growth (\code{"exp"}), log normal random draws (\code{"rlnorm"}), exponential growth with log normal random draws
in the form of process error (\code{"both"}).}

\item{activity.type}{What functional form to use for the underlying activity curves?
Currently supports the Zonneveld model (\code{"zon"}), #' a simple gaussian activity curve (\code{"gauss"}),
and a bivoltine activity curve based on two Gaussian curves (\code{"bivolt"}).}

\item{sample.type}{Distribution to use when simulating censuses. The mean of this distribution
is determined by the underlying activity curve, and this distribution (and associated parameters)
determine the "sampling error" (doesn't need to strictly represent error in simulated observers).
Currently supported: Poisson (\code{"pois"}), negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}).
"nb" and "zinb" will require additional parameters (see details)}

\item{...}{additional arguments for the various timeseries generation functions. See details.}
}
\value{
Data frame of simulated time series. Columns \code{years}, \code{doy}, and \code{count} contain the
the key data: the year and day of simulated observations, and the simulated count of
butterflies observed. \code{abund.true} gives the abundance index of the underlying activity curve
for that year, and \code{act} gives the value of the activity curve for that year and doy
(ie the expectation used to simulate \code{count} from.
\code{onset.true}, \code{median.true}, \code{end.true}, and \code{fp.true} give the phenology metrics calculated
on the underlying activity curve. The \verb{*.true} columns can be used to compare
with estimates of these various metrics. Note that because of random sampling,
the count data will NOT support these estimates in any given instance,
but repeated simulation and analysis using an unbiased method should produce
metrics whose average converges on these \verb{*.true} values. Note that when using
\code{sample.type == "zinb"} option, \code{abund.true} is for the actual activity curve,
but we expect our abundance estimates to be reduced by the zero inflation controlled by
the "pzeros" argument.
}
\description{
Generate time series
}
\details{
For all the additional arguments needed for the various component functions,
see \code{?abund_generator}, \code{activity_gen}, and \code{activity_sampler}. Unfortunately it is not
possible to import details from multiple functions.
}
\examples{
timeseries_generator(years = 1990:2000,
  doy.samples = seq(105,160, by = 7),
  abund.type = "exp",
  activity.type = "gauss",
  sample.type = "pois",
  growth.rate = -0.12,
  init.size = 500,
  act.mean = 130,
  act.sd = 15)
}
