% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_truepheno.R
\name{timeseries_truepheno}
\alias{timeseries_truepheno}
\title{Calculate "true" phenology metrics for a given parameter combination}
\usage{
timeseries_truepheno(activity.type, ...)
}
\arguments{
\item{activity.type}{What functional form to use for the underlying activity curves?
Currently supports the Zonneveld model (\code{"zon"}), #' a simple gaussian activity curve (\code{"gauss"}),
and a bivoltine activity curve based on two Gaussian curves (\code{"bivolt"}).}

\item{...}{further arguments passed to the activity curves. See details.}
}
\value{
1-row data frame of phenology metrics. \code{onset}, \code{median}, and \code{end} give
the days of the 0.1, 0.5, and 0.9 quantile respectively, \code{fp} gives the "flight period",
the number of days between onset and end.
}
\description{
We are simulating distributions that are not changing in phenology across years,
so we only need to estimate the metrics from a single year. This function uses
numeric calculations rather than analytical ones (a) for generalizability to
distributions without closed form solutions for quantiles, and (b) to be exactly comparable
to GAM methods. These phenology metrics are showing what we would calculate if
our GAMs perfectly fit the activity curve.
}
\details{
For the Gaussian distribution, the additional parameters are \code{act.mean} and \code{act.sd},
defining the mean and standard deviation of the associated Guassian distribution.

For the bivoltine distribution, the additional parameters are expanded to define the shape of
two activity curves, with \code{act.mean1} and \code{act.mean2} giving the means of the two gaussians,
and \code{act.sd1} and \code{act.sd2} giving the standard deviations of the two gaussians. Additionally,
\code{rel.size2} gives the size of thes second peak relative to the first. \code{rel.size = 1/2} means the
second peak has half the size of the first, \code{rel.size = 10} means the second peak has 10x the size, etc.
\code{rel.size} must be positive. Note that the code normalizes the sum of the gaussians, so rel.size will
not change the overall abundance.

For the Zonneveld model,
peak emergence time (\code{zon.theta}), spread in emerge times (\code{beta}), and death rate (\code{alpha}).
These parameters
and the activity curve are taken from Zonneveld et al. 2003. For normal use of this
model,
we would have to define initial number of larvae eclosing. However, here we set it such
that the resulting abundance index equals the provided abundance index.
Since area under curve = N / alpha, we set N = abundance index * alpha.
}
\examples{
out = timeseries_truepheno(
  activity.type = "gauss",
  act.mean = 130,
  act.sd = 15)

}
