% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fitter.R
\name{trend_aggregator}
\alias{trend_aggregator}
\title{Fit trend data and aggregate with other information}
\usage{
trend_aggregator(
  dat.summary,
  gam.args,
  timeseries,
  nobs.min,
  nnzero.min,
  nyear.min,
  bound.reasonable.rel = T,
  bound.reasonable.abs = T
)
}
\arguments{
\item{dat.summary}{Data frame with phenology and abundance metrics for any number of years.
Generally just feed in \verb{$summary} component of the results of \code{gam_fitter}. Must include \code{year},
\code{n}, \code{nzero}, \code{abund}, \code{onset}, \code{median}, \code{end}, \code{fp}, \code{boundary.reasonable.rel}, \code{boundary.reasonable.abs},
\code{sim.id}, and \code{gam.id}}

\item{timeseries}{}

\item{nobs.min}{Minimum number of total observations to include a year (REAL observations;
anchors are not counted). This is only relevant if years vary in the number of sampling days}

\item{nnzero.min}{Minimum number of non-zero observations to include a year}

\item{bound.reasonable.rel}{Should we only use years with "good" fits as identified
using bount.reasonable.rel? See \code{gam_summarizer} for details.}

\item{bound.reasonable.abs}{Should we only use years with "good" fits as identified
using bount.reasonable.abs? See \code{gam_summarizer} for details.}
}
\value{
Data frame including fitted trend summary (see \code{trend_method()}),
summary of the number of "unreasonable" gam fits, the real underlying trends,
the exclusion methods used, and the gam methods used. Trend terms starting with "true"
reflect the trend estimated from the (numerically estimated) abundance and phenology
metrics of the underlying activity curve. Because of the randomness associated with
sampling (e.g. Poisson or negative binomial counts), we don't expect the data to exactly
reflect this activity curve even if the gam fit + exclusion rules were perfect.
}
\description{
This is the workhouse function for starting with yearly metrics and fitting parameters
and producing a data frame with all the pieces needed to evaluate the methods.
}
\details{
Fill in column-by-column description here.
}
