% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fitter.R
\name{trend_fitter}
\alias{trend_fitter}
\title{Fit the abundance and phenology trends for summary statistics of time series}
\usage{
trend_fitter(dat.filtered)
}
\arguments{
\item{dat.filtered}{Data frame with phenology and abundance metrics for any number of years.
Generally just feed in \verb{$summary} component of the results of \code{gam_fitter}, with or without
applying censorship filters.}
}
\value{
1-row data frame with linear trends across years for each metric. For abundance,
\code{growth.rate} is the trend in log abundance across years. \verb{$nyears} is the number of unique
years provided.
}
\description{
Fit the abundance and phenology trends for summary statistics of time series
}
\examples{
dat.sim = timeseries_generator(years = 1990:2000,
doy.samples = seq(105,160, by = 7),
abund.type = "exp",
activity.type = "gauss",
sample.type = "pois",
growth.rate = -0.12,
init.size = 500,
act.mean = 130,
act.sd = 15)
out = gam_fitter(years.vec = dat.sim$years,
                 doy.vec = dat.sim$doy,
                 count.vec = dat.sim$count,
                 doy.smooth = "cr",
                 doy.knots = 5,
                 years.smooth = "cr",
                 anchor.flag = FALSE
)
fit_plotter(dat.timeseries = dat.sim,
            dat.fitted = out$dat.fitted,
            activity.curve = out$activity.curve)
trend_fitter(out$summary)
}
