% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fitter.R
\name{trend_method}
\alias{trend_method}
\title{Calculate trends (and diagnostic metrics) for a given censorship method}
\usage{
trend_method(
  dat.summary,
  nobs.min,
  nnzero.min,
  nyear.min,
  bound.reasonable.rel = F,
  bound.reasonable.abs = F
)
}
\arguments{
\item{dat.summary}{Data frame with phenology and abundance metrics for any number of years.
Generally just feed in \verb{$summary} component of the results of \code{gam_fitter}. Must include \code{year},
\code{n}, \code{nzero}, \code{abund}, \code{onset}, \code{median}, \code{end}, \code{fp}, \code{boundary.reasonable.rel}, \code{boundary.reasonable.abs},
\code{sim.id}, and \code{gam.id}}

\item{nobs.min}{Minimum number of total observations to include a year (REAL observations;
anchors are not counted). This is only relevant if years vary in the number of sampling days.
Note that in the current iteration with constant number of days sampled (the vector \code{doy.samples}),
this should be no greater than the length of \code{doy.samples}.}

\item{nnzero.min}{Minimum number of non-zero observations to include a year. This should be
no greater than the length of doy.samples.}

\item{nyear.min}{Minimum number of years to use when fitting, after censoring years
with insufficient data. If fewer years are provided trends are returned as NA. This value
should be no greater than the length of the \code{years} argument, or no simulations will meet
this criterion.}

\item{bound.reasonable.rel}{Should we only use years with "good" fits as identified
using bound.reasonable.rel? See \code{gam_summarizer} for details.}

\item{bound.reasonable.abs}{Should we only use years with "good" fits as identified
using bound.reasonable.abs? See \code{gam_summarizer} for details.}
}
\value{
Data frame with estimated trends (\code{growth.rate}, \code{median}, \code{onset},
\code{end}, \code{fp}), the number of years used to estimate the trends (\code{nyears}),
the filtering criterion used to censor years before fitting trends (arguments for
this function), and the original number of years before censoring (\code{nyear.original}))
}
\description{
Applies to all sim.id and gam.id separately
}
\examples{
set.seed(10)
}
