% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly_gam_sim.R
\name{butterfly_gam_sim}
\alias{butterfly_gam_sim}
\title{Run full simulation}
\usage{
butterfly_gam_sim(
  sim.name,
  path,
  nsims,
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  sim.parms,
  gam.args,
  nobs.min,
  nnzero.min,
  nyear.min,
  nthreads = NULL,
  bound.reasonable.rel = F,
  bound.reasonable.abs = F,
  append = FALSE,
  timeseries.do = TRUE,
  gam.do = TRUE
)
}
\arguments{
\item{sim.name}{Name for saving simulations of this combination of parameters}

\item{path}{File path to save simulated data and results into}

\item{nsims}{Number of simulations to carry out}

\item{years}{Numeric vector of years for which which to simulate data}

\item{doy.samples}{Numeric vector of day of years for which to simulate censusing.}

\item{abund.type}{What type of variation in abundance to model. Currently supported: deterministic
exponential growth (\code{"exp"}), log normal random draws (\code{"rlnorm"}), exponential growth with log normal random draws
in the form of process error (\code{"both"}).}

\item{activity.type}{What functional form to use for the underlying activity curves?
Currently supports the Zonneveld model (\code{"zon"}), #' a simple gaussian activity curve (\code{"gauss"}),
and a bivoltine activity curve based on two Gaussian curves (\code{"bivolt"}).}

\item{sample.type}{Distribution to use when simulating censuses. The mean of this distribution
is determined by the underlying activity curve, and this distribution (and associated parameters)
determine the "sampling error" (doesn't need to strictly represent error in simulated observers).
Currently supported: Poisson (\code{"pois"}), negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}).
"nb" and "zinb" will require additional parameters (see details)}

\item{sim.parms}{list with named components matching each of the required additional
parameters for the selected abundance type, activity type, and sample type.
See details for full list of parameter names. Each sim.parm component can either take
a single value (every simulation will use that value), or vector of the minimum and
maximum desired values (each simulation will use a different value drawn from a uniform
with those bounds).}

\item{gam.args}{Data frame, each row corresponding to a different set of gam method arguments.
Must include columns \verb{$doy.smooth}, \verb{$doy.knots},\verb{$years.smooth},\verb{$anchor.flag}, \verb{$anchor.dist}.
See \code{gam_fitter()} for details. \verb{$anchor.dist} can be any int when \verb{$anchor.flag==FALSE};
I recommend using a large negative number to make it easy to catch mistakes.}

\item{nobs.min}{Minimum number of total observations to include a year (REAL observations;
anchors are not counted). This is only relevant if years vary in the number of sampling days.
Note that in the current iteration with constant number of days sampled (the vector \code{doy.samples}),
this should be no greater than the length of \code{doy.samples}.}

\item{nnzero.min}{Minimum number of non-zero observations to include a year. This should be
no greater than the length of doy.samples.}

\item{nyear.min}{Minimum number of years to use when fitting, after censoring years
with insufficient data. If fewer years are provided trends are returned as NA. This value
should be no greater than the length of the \code{years} argument, or no simulations will meet
this criterion.}

\item{nthreads}{Defaults to NULL. If an integer, gam_fitall_gen will call a parallelized version of
the gam fitting function (\code{gam_fitall_par()}), which will use \code{nthread} threads to fit gams in parallel.}

\item{bound.reasonable.rel}{Should we only use years with "good" fits as identified
using bound.reasonable.rel? See \code{gam_summarizer} for details.}

\item{bound.reasonable.abs}{Should we only use years with "good" fits as identified
using bound.reasonable.abs? See \code{gam_summarizer} for details.}

\item{append}{For debugging purposes only. Leave at default value of \code{FALSE}.}

\item{timeseries.do}{For debugging purposes only. Leave at default value of \code{TRUE}}

\item{gam.do}{For debugging purposes only. Leave at default value of \code{TRUE}}
}
\value{
No object is returned, but a series of files are generated in \code{path/sim.name},
which can be easily loaded with \code{butterfly_load}.
}
\description{
This is the main user function for this package. \code{butterfly_gam_sim()} takes timeseries parameters,
gam-fitting parameters, and censoring parameters, generates appropriate time series, fits
those time series with appropriate gams, estimates phenology metrics and abundance index
for each year, then looks for linear (log-linear in the case of abundance) trends across
years after applying the appropriate filters of \code{nobs.min}, \code{nnzero.min}, \code{nyear.min}.
Results are saved with a consistent file structure in the \code{sim.name} folder of the
\code{path} directory. For easy access to these simulated files, see \code{butterfly_load()}.
}
\details{
To allow simulations to capture a range of possible population patterns, activity curves and sampling
distributions, \code{sim.parms} has variable requirements based on \code{sample.type}, \code{abund.type}, and \code{activity.type}.
For \code{sample_type = "nb"}, theta must be included in \code{sim.parms}. For \code{sample_type = "zinb"},
\code{theta} and \code{pzero} must be included in \code{sim.parms}.#' See \code{activity_sampler()} for details.

For \code{activity.type = "gaus"}, \code{sim.parms} must include \code{act.mean} (peak day) and \code{act.sd} (peak width).
For \code{"zon"}, \code{sim.parms} must include \code{zon.theta} (peak day relative to day 0), \code{t0} (the "initial" day of the zonneveld model,
which is the final day that populations are 0), \code{beta} (spread), and \code{alpha} (death rate).
For "bivolt", \code{sim.parms} must include \code{act.mean1} and \code{act.mean2} (peak days), \code{act.sd1} and
\code{act.sd2} (peak widths), and \code{rel.size} (ratio of peak sizes). See \code{activity_gen()} for details.

\code{sim.parms} must also contain information about the underlying population. For \code{abund.type = "exp"} (exponential growth),
this means the initial abundance index \code{init.size} and the growth rate \code{growth.rate}. For \code{abund.type = "rlnorm"},
(log-nromal random draws), this means \code{meanlog} and \code{sdlog}. See \code{abund_generator} for details.
}
\examples{
#  path.res = "G:/Dropbox/academia/research projects/butterfly-gam-sims/4_res"
#  gam.args = data.frame(doy.smooth = c("cc", "cc", "cc"),
#                       doy.knots = c(5,3, 5),
#                       years.smooth = c("cr", "cr", "cr"),
#                       anchor.flag = c(TRUE, TRUE, FALSE),
#                       anchor.dist = c(10, 10, 10))
# sim.parms = list(growth.rate = -0.12,
#                  init.size = 500,
#                  act.mean = 130,
#                  act.sd = 15,
#                  theta = 5)
# butterfly_gam_sim(sim.name = "full-kaboodle",
#                   path = path.res,
#                   nsims = 10,
#                   years = 1990:2010,
#                   doy.samples = seq(0, 365, by = .1),
#                   abund.type = "exp",
#                   activity.type = "gauss",
#                   sample.type = "pois",
#                   sim.parms = sim.parms,
#                   gam.args = gam.args,
#                   nobs.min = 5,
#                   nnzero.min = 5,
#                   nyear.min = 3)
# res = butterfly_load(sim.name = "full-kaboodle",
#                      path = path.res)
}
