% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_summarizer.R
\name{gam_summarize_all}
\alias{gam_summarize_all}
\title{Estimate metrics for full activity curve}
\usage{
gam_summarize_all(activity.curve, ...)
}
\arguments{
\item{activity.curve}{data frame with \verb{$years}, \verb{$doy}, and \verb{$act}, typically generated
withing \code{gam_fitter()}}

\item{...}{\code{"bounds.reasonable"}, "\verb{bounds.thresh.rel"}, and \code{"bounds.thresh.abs"} can
be specified here to pass on to \code{gam_summarizer()}.}
}
\value{
Data frame with summary metrics for each year. see \code{gam_summarizer()} for details.

Helper function to summarize multi-year activity curve dataframe with \code{gam_summarizer}
which (CURRENTLY) fits one year at a time. This function makes it easier to
update code to use an optimized \code{gam_summarizer} if I get there.
Note: this is a bit fragile, and sensitive to changes in what \code{gam_summarizer} outputs.
Not sure if that's worth working around though. Leaving for now.
}
\description{
Estimate metrics for full activity curve
}
