% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_examples.R
\name{timeseries_examples}
\alias{timeseries_examples}
\title{Plot example timeseries}
\usage{
timeseries_examples(
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  xlim = c(0, 365),
  ...
)
}
\arguments{
\item{years}{Years  for which to simulate data}

\item{doy.samples}{Days of year to simulate observation events.}

\item{abund.type}{What type of variation in abundance to model. Currently supported:
exponential growth (\code{"exp"}), log normal variation (\code{"rlnorm"}).}

\item{activity.type}{What type activity curve to model? Currently supports the
Zonneveld model ("Zonneveld"),
a simple gaussian activity curve ("Gauss"), and a bivoltine activity curve based
on two Gaussian curves ("bivoltine").}

\item{sample.type}{Name of sampling distribution. Currently supported: Poisson (\code{"pois"}),
negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}). "nb" and "zinb"
will require additional parameters.}

\item{xlim}{vector of limits for plotting}

\item{...}{additional arguments for the various timeseries generation functions. See details.}
}
\value{
ggplot object
}
\description{
Plot example timeseries
}
\details{
For all the additional arguments needed for the various component functions,
see \code{?abund_generator}, \code{activity_gen}, and \code{activity_sampler}. Unfortunately it is not
possible to import details from multiple functions.
}
\examples{
timeseries_examples(years = 1990:2000,
  doy.samples = seq(105,160, by = 7),
  abund.type = "exp",
  activity.type = "Gauss",
  sample.type = "pois",
  growth.rate = -0.12,
  init.size = 500,
  act.mean = 130,
  act.sd = 15)
}
