% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_generator.R
\name{timeseries_generator}
\alias{timeseries_generator}
\title{Generate time series}
\usage{
timeseries_generator(
  years,
  doy.samples,
  abund.type,
  activity.type,
  sample.type,
  ...
)
}
\arguments{
\item{years}{Years  for which to simulate data}

\item{doy.samples}{Days of year to simulate observation events.}

\item{abund.type}{What type of variation in abundance to model. Currently supported:
exponential growth (\code{"exp"}), log normal variation (\code{"rlnorm"}).}

\item{activity.type}{What type activity curve to model? Currently supports the
Zonneveld model (\code{"zon"}),
a simple gaussian activity curve (\code{"gauss"}), and a bivoltine activity curve based
on two Gaussian curves (\code{"bivolt"}).}

\item{sample.type}{Name of sampling distribution. Currently supported: Poisson (\code{"pois"}),
negative binomial (\code{"nb"}), and zero-inflated negative binomial (\code{"zinb"}). "nb" and "zinb"
will require additional parameters.}

\item{...}{additional arguments for the various timeseries generation functions. See details.}
}
\value{
Data frame of simulated time series. Columns \code{years}, \code{doy}, and \code{count} contain the
the key data: the year and day of simulated observations, and the simulated count of
butterflies observed. \code{abund.true} gives the abundance index of the underlying activity curve
for that year, and \code{act} gives the value of the activity curve for that year and doy
(ie the expectation used to simulate \code{count} from.
\code{onset.true}, \code{median.true}, \code{end.true}, and \code{fp.true} give the phenology metrics calculated
on the underlying activity curve. The \verb{*.true} columns can be used to compare
with estimates of these various metrics. Note that because of random sampling,
the count data will NOT support these estimates in any given instance,
but repeated simulation and analysis using an unbiased method should produce
metrics whose average converges on these \verb{*.true} values. Note that when using
\code{sample.type == "zinb"} option, \code{abund.true} is for the actual activity curve,
but we expect our abundance estimates to be reduced by the zero inflation controlled by
the "pzeros" argument.
}
\description{
Generate time series
}
\details{
For all the additional arguments needed for the various component functions,
see \code{?abund_generator}, \code{activity_gen}, and \code{activity_sampler}. Unfortunately it is not
possible to import details from multiple functions.
}
\examples{
timeseries_generator(years = 1990:2000,
  doy.samples = seq(105,160, by = 7),
  abund.type = "exp",
  activity.type = "gauss",
  sample.type = "pois",
  growth.rate = -0.12,
  init.size = 500,
  act.mean = 130,
  act.sd = 15)
}
