% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fitter.R
\name{trend_aggregator}
\alias{trend_aggregator}
\title{Fit trend data and aggregate with other information}
\usage{
trend_aggregator(
  dat.summary,
  gam.args,
  timeseries,
  nobs.min,
  nnzero.min,
  nyear.min,
  bound.reasonable.rel = T,
  bound.reasonable.abs = T
)
}
\arguments{
\item{dat.summary}{Data frame with phenology and abundance metrics for any number of years.
Generally just feed in \verb{$summary} component of the results of \code{gam_fitter}. Must include \code{year},
\code{n}, \code{nzero}, \code{abund}, \code{onset}, \code{median}, \code{end}, \code{fp}, \code{boundary.reasonable.rel}, \code{boundary.reasonable.abs},
\code{sim.id}, and \code{gam.id}}

\item{gam.args}{Data frame, each row corresponding to a different set of gam method arguments.
Must include columns \verb{$doy.smooth}, \verb{$doy.knots},\verb{$years.smooth},\verb{$anchor.flag}, \verb{$anchor.dist}.
See \code{gam_fitter()} for details. \verb{$anchor.dist} can be any int when \verb{$anchor.flag==FALSE};
I recommend using a large negative number to make it easy to catch mistakes.}

\item{timeseries}{Collection of timeseries, including \verb{$years}, \verb{$doy}, \verb{$count}, and \verb{$sim.id}.
Typically the \verb{$timeseries} output of \code{timeseries_sim()} or \code{timeseries_load()}. If speed is
of particular concern, can pass the trimmed data frame (\code{timeseries_sim()} generates
many more columns.)}

\item{nobs.min}{Minimum number of total observations to include a year (REAL observations;
anchors are not counted). This is only relevant if years vary in the number of sampling days.
Note that in the current iteration with constant number of days sampled (the vector \code{doy.samples}),
this should be no greater than the length of \code{doy.samples}.}

\item{nnzero.min}{Minimum number of non-zero observations to include a year. This should be
no greater than the length of doy.samples.}

\item{nyear.min}{Minimum number of years to use when fitting, after censoring years
with insufficient data. If fewer years are provided trends are returned as NA. This value
should be no greater than the length of the \code{years} argument, or no simulations will meet
this criterion.}

\item{bound.reasonable.rel}{Should we only use years with "good" fits as identified
using bound.reasonable.rel? See \code{gam_summarizer} for details.}

\item{bound.reasonable.abs}{Should we only use years with "good" fits as identified
using bound.reasonable.abs? See \code{gam_summarizer} for details.}
}
\value{
Data frame including fitted trend summary (see \code{trend_method()}),
summary of the number of "unreasonable" gam fits, the real underlying trends,
the exclusion methods used, and the gam methods used. Trend terms starting with "true"
reflect the trend estimated from the (numerically estimated) abundance and phenology
metrics of the underlying activity curve. Because of the randomness associated with
sampling (e.g. Poisson or negative binomial counts), we don't expect the data to exactly
reflect this activity curve even if the gam fit + exclusion rules were perfect.
}
\description{
This is the workhouse function for starting with yearly metrics and fitting parameters
and producing a data frame with all the pieces needed to evaluate the methods.
}
\details{
Fill in column-by-column description here.
}
